refactor(Algolia CommandMenu): Refactored to basic TS
- move CommandMenu component to search/CommandMenu directory for test readiness
- move AlgoliaSearch component to search/AlgoliaSearch directory test readiness
- extract searchClient to separate file in search directory
- minor style updates for ARIA

Close T-12339
