+++ src/components/shared/search/searchClient.ts
@@ -1,0 +1,6 @@
+import algoliasearch from "algoliasearch/lite";
+
+export const searchClient = algoliasearch(
+    "U5XNUAICIA",
+    "5086892b29c9c6a9408401d4dc5313fd"
+);
